pyproject.toml
[project]
name = "aiac-be"
version = "1.0.0"
description = "CDP Document Analysis Backend - Analisi e confronto documenti normativi via LLM"
authors = [{ name = "CDP Team" }]
readme = "README.md"
requires-python = ">=3.11"

[project.scripts]
# CLI principale
lex-cli = "lex_package.cli:main"

[tool.poetry]
packages = [
    { include = "lex_package", from = "src/be/src" },
    { include = "core", from = "src" },
    { include = "utils", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"

# === Core LLM & AI ===
langchain = "^0.3.18"
langchain-community = "^0.3.15"
langchain-core = "^0.3.34"
langchain-openai = "^0.3.4"
langchain-text-splitters = "^0.3.6"
langgraph = "^0.2.65"
langgraph-checkpoint = "^2.0.10"
langgraph-sdk = "^0.1.51"
openai = "^1.55.0"
tiktoken = "^0.8.0"

# === Azure Services ===
azure-storage-blob = "^12.24.0"
azure-search-documents = "11.5.2"
azure-ai-documentintelligence = "^1.0.0"
azure-ai-formrecognizer = "^3.3.3"
azure-identity = "^1.21.0"
azure-cosmos = "^4.9.0"
azure-keyvault-secrets = "^4.9.0"
azure-core = "^1.32.0"
langgraph-checkpoint-cosmosdb = "^0.2.3"

# === Web Framework ===
fastapi = {extras = ["standard"], version = "^0.115.0"}
uvicorn = "^0.34.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.7.1"

# === PDF Processing ===
PyPDF2 = "^3.0.1"
PyMuPDF = "^1.24.13"
pdfminer-six = "^20240706"

# === Data Processing ===
pandas = "^2.1.4"
numpy = "1.26.4"
scipy = "^1.15.1"
openpyxl = "^3.1.5"
XlsxWriter = "^3.2.0"

# === NLP & Text Processing ===
spacy = "^3.7.5"
python-Levenshtein = "^0.26.1"
fuzzywuzzy = "^0.18.0"
sentence-transformers = "^3.3.1"
transformers = "^4.46.1"
torch = "^2.5.1"

# === Utilities ===
python-dotenv = "^1.0.1"
tqdm = "^4.66.6"
aiofiles = "^24.1.0"
httpx = "^0.28.1"
requests = "^2.32.3"
markdown2 = "^2.5.1"
orjson = "^3.10.15"
tenacity = "^9.0.0"
loguru = "^0.7.2"

# === Monitoring (Azure) ===
azure-monitor-opentelemetry = "^1.6.5"
opentelemetry-api = ">=1.30.0"
opentelemetry-sdk = ">=1.30.0"
# opentelemetry-instrumentation-* managed by azure-monitor-opentelemetry

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
ipywidgets = "^8.0.4"
langchain-cli = "^0.0.36"

[tool.poetry.group.azure-functions]
optional = true

[tool.poetry.group.azure-functions.dependencies]
azure-functions = "^1.21.0"
azure-functions-durable = "^1.2.9"
azurefunctions-extensions-bindings-blob = "^1.0.0"
azurefunctions-extensions-http-fastapi = "^1.0.0"
scikit-learn = "^1.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/be/src/lex_package", "src/be/azure-durable-function/tests"]
pythonpath = ["src", "src/be/src", "src/be/azure-durable-function"]
